Move typings folder
