+++ package.json
@@ -201,7 +201,7 @@
     "vscode:prepublish": "tsc -p ./",
     "compile": "tsc -watch -p ./",
     "postinstall": "node ./node_modules/vscode/bin/install",
-    "tslint": "tslint -t verbose src/**/*.ts",
+    "tslint": "tslint -t verbose src/**/*.ts --exclude src/**/*.d.ts",
     "test": "node ./node_modules/vscode/bin/test"
   },
   "devDependencies": {
@@ -210,6 +210,7 @@
     "@types/mocha": "^2.2.32",
     "@types/node": "^6.0.40",
     "@types/request-promise": "^4.1.41",
+    "@types/strip-json-comments": "0.0.30",
     "azure-arm-resource": "^3.1.1-preview",
     "mocha": "^2.3.3",
     "tslint": "^5.7.0",
