+++ src/common/constants.ts
@@ -29,7 +29,7 @@ export class Constants {
     public static ACR_MODULE = "Existing Module on Azure Container Registry";
     public static EXISTING_MODULE = "Existing Module";
     public static LANGUAGE_CSHARP_DESCRIPTION = "Use Azure IoT C# SDK to build a module";
-    public static LANGUAGE_NODE_DESCRIPTION= "Use Azure IoT Node.js SDK to build a module";
+    public static LANGUAGE_NODE_DESCRIPTION = "Use Azure IoT Node.js SDK to build a module";
     public static LANGUAGE_PYTHON_DESCRIPTION = "Use Azure IoT Python SDK to build a module";
     public static CSHARP_FUNCTION_DESCRIPTION = "Create an Azure Function and deploy to IoT Edge";
     public static ACR_MODULE_DESCRIPTION = "Import an exsting module image from Azure Container Registry";
